make code dynamic
